+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Don't clean up stale storage data in railtie
+- Add enabled configuration option and force parameter for per-request profiling
 
 ## [0.5.0] - 2025-09-21